# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: SkillScan API
  version: "1.1"
  description: |
    SkillScan 团队使用的统一接口。
    包含健康检查、GitHub 安装绑定、报告上传、分享页、扫描记录、审批和审计。
servers:
  - url: https://api.skillscan.example.com
    description: Replace with your real base URL
paths:
...
  /github/install/start:
    post:
      summary: Start GitHub App install flow
      security:
        - sessionCookie: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [install_url]
                properties:
                  install_url:
                    type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /github/install/complete:
    post:
      summary: Complete GitHub App install binding
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [installation_id, state]
              properties:
                installation_id:
                  type: integer
                setup_action:
                  type: string
                state:
                  type: string
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request
        "401":
          description: Not authenticated
        "403":
          description: Invalid state or tenant conflict
        "409":
          description: Installation already linked
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - not_authenticated
            - state_invalid_or_expired
            - tenant_owned_by_another_user
            - installation_already_linked
            - repo_not_enrolled
            - quota_exceeded
            - rate_limited
            - invalid_signature
            - identity_not_allowed
            - report_digest_mismatch
            - schema_invalid
            - internal_error
...